refactor: Playable 인터페이스를 추상클래스로 수정
